moves createNewUser logic out of the server and into a UsersController
